requirements.txt
# Core Web Framework (Pinned to a stable, compatible version)
fastapi==0.104.1
uvicorn[standard]
jinja2
python-multipart

# Database and Data Validation
sqlmodel
psycopg2-binary

# Asynchronous Tasks
celery
redis

# LLM Abstraction
litellm

# Authentication & Security (Pinned to versions compatible with older Python)
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
python-jose[cryptography]

# Environment & Utilities
python-dotenv
aiofiles

# Windows compatibility for Celery
gevent

# Explicitly require Pydantic v2 for FastAPI compatibility
pydantic>=2.0.0,<3.0.0

# Provider-specific libraries for LiteLLM
google-generativeai
anthropics
openai
